Add `-y, --yes` and `-n, --no` flags

+++ src/evaluator.rs
@@ -3,7 +3,7 @@ use std::{ffi::OsStr, fs, io::Write, path::PathBuf};
 use colored::Colorize;
 
 use crate::{
-    cli::{Command, CommandKind},
+    cli::{Command, CommandKind, Flags},
     compressors::{
         BzipCompressor, Compressor, Entry, GzipCompressor, LzmaCompressor, TarCompressor,
         ZipCompressor,
@@ -12,8 +12,8 @@ use crate::{
         BzipDecompressor, DecompressionResult, Decompressor, GzipDecompressor, LzmaDecompressor,
         TarDecompressor, ZipDecompressor,
     },
-    extension::{CompressionFormat, Extension},
     dialogs::Confirmation,
+    extension::{CompressionFormat, Extension},
     file::File,
     utils,
 };
@@ -98,13 +98,13 @@ impl Evaluator {
         Ok((first_decompressor, second_decompressor))
     }
 
-    // todo: move this folder into decompressors/ later on
     fn decompress_file_in_memory(
         bytes: Vec<u8>,
         file_path: PathBuf,
         decompressor: Option<Box<dyn Decompressor>>,
         output_file: &Option<File>,
         extension: Option<Extension>,
+        flags: Flags,
     ) -> crate::Result<()> {
         let output_file_path = utils::get_destination_path(output_file);
 
@@ -113,7 +113,7 @@ impl Evaluator {
             .unwrap_or_else(|| output_file_path.as_os_str());
 
         if filename == "." {
-            // I believe this is only possible when the supplied inout has a name
+            // I believe this is only possible when the supplied input has a name
             // of the sort `.tar` or `.zip' and no output has been supplied.
             filename = OsStr::new("ouch-output");
         }
@@ -127,6 +127,14 @@ impl Evaluator {
                 // Therefore, we'll save what we have in memory into a file.
                 println!("{}: saving to {:?}.", "info".yellow(), filename);
 
+                if filename.exists() {
+                    let confirm =
+                        Confirmation::new("Do you want to overwrite 'FILE'?", Some("FILE"));
+                    if !utils::permission_for_overwriting(&filename, flags, &confirm)? {
+                        return Ok(());
+                    }
+                }
+
                 let mut f = fs::File::create(output_file_path.join(filename))?;
                 f.write_all(&bytes)?;
                 return Ok(());
@@ -139,7 +147,7 @@ impl Evaluator {
             extension,
         };
 
-        let decompression_result = decompressor.decompress(file, output_file)?;
+        let decompression_result = decompressor.decompress(file, output_file, flags)?;
         if let DecompressionResult::FileInMemory(_) = decompression_result {
             // Should not be reachable.
             unreachable!();
@@ -148,19 +156,18 @@ impl Evaluator {
         Ok(())
     }
 
-    fn compress_files(files: Vec<PathBuf>, mut output: File) -> crate::Result<()> {
+    fn compress_files(files: Vec<PathBuf>, mut output: File, flags: Flags) -> crate::Result<()> {
         let confirm = Confirmation::new("Do you want to overwrite 'FILE'?", Some("FILE"));
         let (first_compressor, second_compressor) = Self::get_compressor(&output)?;
 
+        // TODO: use -y and -n here
         let output_path = output.path.clone();
         if output_path.exists() {
-            let output_path_str = &*output_path.to_string_lossy();
-            if !confirm.ask(Some(output_path_str))? {
+            if !utils::permission_for_overwriting(&output_path, flags, &confirm)? {
                 // The user does not want to overwrite the file
                 return Ok(());
             }
         }
-
 
         let bytes = match first_compressor {
             Some(first_compressor) => {
@@ -188,14 +195,13 @@ impl Evaluator {
         Ok(())
     }
 
-    fn decompress_file(file: File, output: &Option<File>) -> crate::Result<()> {
-        // let output_file = &command.output;
+    fn decompress_file(file: File, output: &Option<File>, flags: Flags) -> crate::Result<()> {
         let (first_decompressor, second_decompressor) = Self::get_decompressor(&file)?;
 
         let file_path = file.path.clone();
         let extension = file.extension.clone();
 
-        let decompression_result = second_decompressor.decompress(file, output)?;
+        let decompression_result = second_decompressor.decompress(file, output, flags)?;
 
         match decompression_result {
             DecompressionResult::FileInMemory(bytes) => {
@@ -207,6 +213,7 @@ impl Evaluator {
                     first_decompressor,
                     output,
                     extension,
+                    flags,
                 )?;
             }
             DecompressionResult::FilesUnpacked(_files) => {
@@ -223,18 +230,18 @@ impl Evaluator {
         Ok(())
     }
 
-    pub fn evaluate(command: Command) -> crate::Result<()> {
+    pub fn evaluate(command: Command, flags: Flags) -> crate::Result<()> {
         let output = command.output.clone();
 
         match command.kind {
             CommandKind::Compression(files_to_compress) => {
                 // Safe to unwrap since output is mandatory for compression
                 let output = output.unwrap();
-                Self::compress_files(files_to_compress, output)?;
+                Self::compress_files(files_to_compress, output, flags)?;
             }
             CommandKind::Decompression(files_to_decompress) => {
                 for file in files_to_decompress {
-                    Self::decompress_file(file, &output)?;
+                    Self::decompress_file(file, &output, flags)?;
                 }
             }
         }
